Tweak Convert's code framework, for ease of enhancement
[1] Split the converter classes in Convert into separate files
[2] Add the converter field to ConfigFiled, used to specify the converter class of the field
