+++ eventmesh-common/src/main/java/org/apache/eventmesh/common/config/convert/converter/IPAddressConverter.java
@@ -15,32 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.common.config;
+package org.apache.eventmesh.common.config.convert.converter;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import org.apache.eventmesh.common.config.convert.ConvertInfo;
+import org.apache.eventmesh.common.config.convert.ConvertValue;
+
+import inet.ipaddr.AddressStringException;
+import inet.ipaddr.IPAddress;
+import inet.ipaddr.IPAddressString;
 
 /**
- * Record information about the configuration class to be converted
+ * Config field conversion class for IPAddress
  */
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD, ElementType.TYPE})
-public @interface Config {
+public class IPAddressConverter implements ConvertValue<IPAddress> {
 
-    String field() default "";
-
-    String path() default "";
-
-    String prefix() default "";
-
-    String hump() default ".";
-
-    boolean removePrefix() default true;
-
-    boolean monitor() default false;
+    @Override
+    public IPAddress convert(ConvertInfo convertInfo) {
+        try {
+            return new IPAddressString((String) convertInfo.getValue()).toAddress();
+        } catch (AddressStringException e) {
+            throw new RuntimeException(e);
+        }
+    }
 }
-
-
-
