+++ eventmesh-common/src/main/java/org/apache/eventmesh/common/config/convert/converter/StringConverter.java
@@ -15,32 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.common.config;
+package org.apache.eventmesh.common.config.convert.converter;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import org.apache.eventmesh.common.config.convert.ConvertInfo;
+import org.apache.eventmesh.common.config.convert.ConvertValue;
 
 /**
- * Record information about the configuration class to be converted
+ * Config field conversion class for String
  */
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD, ElementType.TYPE})
-public @interface Config {
+public class StringConverter  implements ConvertValue<String> {
 
-    String field() default "";
-
-    String path() default "";
-
-    String prefix() default "";
-
-    String hump() default ".";
-
-    boolean removePrefix() default true;
-
-    boolean monitor() default false;
+    @Override
+    public String convert(ConvertInfo convertInfo) {
+        return (String) convertInfo.getValue();
+    }
 }
-
-
-
